refactor: split storage internals from common
