+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["api", "cli", "common", "lite", "sdk"]
+members = ["api", "cli", "common", "lite", "resource-spec", "sdk", "storage"]
 
 [workspace.package]
 edition = "2024"
@@ -53,7 +53,9 @@ rustls = { version = "0.23", default-features = false, features = ["logging", "s
 s2-api = { path = "api", version = "0.30" }
 s2-common = { path = "common", version = "0.38" }
 s2-lite = { path = "lite", version = "0.36" }
+s2-resource-spec = { path = "resource-spec", version = "0.1" }
 s2-sdk = { path = "sdk", version = "0.31" }
+s2-storage = { path = "storage", version = "0.1" }
 schemars = "1.2"
 secrecy = "0.10.3"
 serde = "1.0"
