+++ common/Cargo.toml
@@ -11,28 +11,21 @@ keywords = ["s2", "durable", "streams", "log", "api"]
 [features]
 axum = ["dep:axum"]
 clap = ["dep:clap"]
-resource_spec = ["dep:humantime", "dep:schemars", "dep:serde_json"]
 rkyv = ["dep:rkyv", "compact_str/rkyv"]
 utoipa = ["dep:utoipa"]
 
 [dependencies]
-aegis = { workspace = true }
-aes-gcm = { workspace = true, features = ["zeroize"] }
 axum = { workspace = true, optional = true }
 base64ct = { workspace = true, features = ["alloc"] }
-blake3 = { workspace = true }
 bytes = { workspace = true }
 clap = { workspace = true, optional = true, features = ["derive"] }
 compact_str = { workspace = true, features = ["serde"] }
 enumset = { workspace = true }
 http = { workspace = true }
-humantime = { workspace = true, optional = true }
 rand = { workspace = true }
 rkyv = { workspace = true, optional = true }
-schemars = { workspace = true, optional = true }
 secrecy = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
-serde_json = { workspace = true, optional = true }
 strum = { workspace = true, features = ["derive"] }
 thiserror = { workspace = true }
 time = { workspace = true }
