+++ common/src/lib.rs
@@ -1,6 +1,5 @@
 //! Common types and utilities shared across S2 crates.
 
-pub mod bash;
 pub mod caps;
 pub mod deep_size;
 pub mod encryption;
@@ -8,6 +7,4 @@ pub mod http;
 pub mod maybe;
 pub mod read_extent;
 pub mod record;
-#[cfg(feature = "resource_spec")]
-pub mod resource_spec;
 pub mod types;
