+++ lite/src/backend/kv/basin_meta.rs
@@ -1,13 +1,11 @@
 use std::ops::Range;
 
 use bytes::{BufMut, Bytes, BytesMut};
-use s2_common::{
-    bash::Bash,
-    types::{
-        basin::{BasinName, BasinNamePrefix, BasinNameStartAfter},
-        config::BasinConfig,
-    },
+use s2_common::types::{
+    basin::{BasinName, BasinNamePrefix, BasinNameStartAfter},
+    config::BasinConfig,
 };
+use s2_storage::bash::Bash;
 use serde::{Deserialize, Serialize};
 use time::OffsetDateTime;
 
@@ -119,13 +117,11 @@ mod tests {
 
     use bytes::Bytes;
     use proptest::prelude::*;
-    use s2_common::{
-        bash::Bash,
-        types::{
-            basin::{BasinName, BasinNamePrefix, BasinNameStartAfter},
-            config::{BasinConfig, OptionalDeleteOnEmptyConfig, OptionalStreamConfig},
-        },
+    use s2_common::types::{
+        basin::{BasinName, BasinNamePrefix, BasinNameStartAfter},
+        config::{BasinConfig, OptionalDeleteOnEmptyConfig, OptionalStreamConfig},
     };
+    use s2_storage::bash::Bash;
     use time::OffsetDateTime;
 
     use crate::backend::kv::{
