+++ storage/src/bash.rs
