+++ storage/src/lib.rs
@@ -1,0 +1,5 @@
+//! Storage/backend shared code for S2.
+
+pub mod bash;
+pub mod record;
+pub mod types;
